MAJOR: +Schedule Edit //real bad code(( +Cabinets +WASM browser (+JS, but Haze doesn't support it yet) SOME: CDispatcher (IO except browsers) New menus

+++ common/admin/presentation/src/commonMain/kotlin/groups/students/StudentsComponent.kt
@@ -35,7 +35,7 @@ class StudentsComponent(
         storeFactory,
         name = "formListInGroups",
         onItemClick = {
-            onEvent(StudentsStore.Intent.BindStudentToForm(it.id))
+            onEvent(StudentsStore.Intent.BindStudentToForm(it.id.toInt()))
         })
 
     private val studentsStore =
